requirements.txt
flask
mysql-connector-python
python-nmap
scikit-learn
joblib
python-dotenv
